requirements.txt
requests
pydantic
pandas
retrying
jinja2
pytest
redis
PyYAML
black
flake8